Upgraded to Hibernate 3.6.10 and removed the dependeny on: - hibernate-annotations - hibernate-commons-annotations
They finally cleared up the mess with the jars and versions:

In 3.5, everything is released together, with the same version number:
- https://community.jboss.org/wiki/HibernateCompatibilityMatrix
- http://in.relation.to/Bloggers/ClarificationAboutAnnotationsAndEntityManager

In 3.6, they dropped support for JDK 1.4, and they merged the content
of the hibernate-annotations and hibernate-jmx into hibernate-core, as
described in the 3.6 release notes:

http://in.relation.to/Bloggers/HibernateCore360FinalRelease

It looks like hibernate-commons-annotations wasn't necessary anymore
for a longer time already, but I haven't figured that out. Glad it's
gone.